+++ NEWS.md
@@ -1,3 +1,19 @@
+# BuzzardsBay 0.1.0.9004
+
+* Add `bb_options()` for setting and retrieving global parameters which are
+  stored in a private environment, `bbp`.
+* Move auto qc parameters from `[qc_deployment()] and helper functions into
+  global parameters.
+* Add `tide_station_info` dataset containing locations of NOAA tide stations 
+  in Buzzard's bay.
+* Add `inst/rmd/QAQC_report.rmd` with plots and summary information to
+  facilitate quality control and quality assurance.
+* Add `make_deployment_report()` to make QAQC report from data in a deployment
+  directory.
+* `qc_deployment()` now makes the report automatically unless the new `report`
+  argument is set to `FALSE`.
+
+
 # BuzzardsBay 0.1.0.9003
 2024-05-10